django rest/requirements.txt
Django==5.0.2
django-rest-framework==0.1.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
python-decouple==3.8 